SdmGoalImplementationMapper now an interface
