+++ src/api/goal/support/executeSendMessageToSlack.ts
