+++ src/goal/SdmGoalImplementationMapperImpl.ts
@@ -1,69 +1,23 @@
-/*
- * Copyright © 2018 Atomist, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+import { Goal } from "../api/goal/Goal";
+import { IsolatedGoalLauncher } from "../api/goal/support/IsolatedGoalLauncher";
+import {
+    GoalFulfillment,
+    GoalFullfillmentCallback,
+    GoalImplementation,
+    GoalSideEffect,
+    SdmGoalImplementationMapper,
+} from "../api/goal/support/SdmGoalImplementationMapper";
+import { PushListenerInvocation } from "../api/listener/PushListener";
+import { SdmGoal } from "../ingesters/sdmGoalIngester";
 
-import { IsolatedGoalLauncher } from "../../handlers/events/delivery/goals/launchGoal";
-import { SdmGoal } from "../../ingesters/sdmGoalIngester";
-import { InterpretLog } from "../../spi/log/InterpretedLog";
-import { RepoContext } from "../context/SdmContext";
-import { PushListenerInvocation } from "../listener/PushListener";
-import { PushTest } from "../mapping/PushTest";
-import { ExecuteGoalWithLog } from "./ExecuteGoalWithLog";
-import { Goal } from "./Goal";
-
-export type GoalFulfillment = GoalImplementation | GoalSideEffect;
-
-export interface GoalImplementation {
-    implementationName: string;
-    goal: Goal;
-    goalExecutor: ExecuteGoalWithLog;
-    pushTest: PushTest;
-    logInterpreter: InterpretLog;
-}
-
-export function isGoalImplementation(f: GoalFulfillment): f is GoalImplementation {
-    return !!f && !!(f as GoalImplementation).implementationName && true;
-}
-
-export interface GoalSideEffect {
-    sideEffectName: string;
-    goal: Goal;
-    pushTest: PushTest;
-}
-
-export function isSideEffect(f: GoalFulfillment): f is GoalSideEffect {
-    return !!f && (f as GoalSideEffect).sideEffectName && true;
-}
-
-/**
- * Callback to allow changes to the goal before it gets fullfilled.
- *
- * This is useful to add goal specific information to the data field.
- */
-export interface GoalFullfillmentCallback {
-    goal: Goal;
-    callback: (goal: SdmGoal, context: RepoContext) => Promise<SdmGoal>;
-}
-
-export class SdmGoalImplementationMapper {
+export class SdmGoalImplementationMapperImpl implements SdmGoalImplementationMapper {
 
     private readonly implementations: GoalImplementation[] = [];
     private readonly sideEffects: GoalSideEffect[] = [];
     private readonly callbacks: GoalFullfillmentCallback[] = [];
 
-    constructor(private readonly goalLauncher: IsolatedGoalLauncher) {}
+    constructor(private readonly goalLauncher: IsolatedGoalLauncher) {
+    }
 
     public findImplementationBySdmGoal(goal: SdmGoal): GoalImplementation {
         const matchedNames = this.implementations.filter(m =>
