+++ src/goal/common/npmGoals.ts
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-import { IndependentOfEnvironment, ProductionEnvironment, StagingEnvironment } from "../../api/goal/common/environment";
 import { GoalWithPrecondition } from "../../api/goal/Goal";
 import { Goals } from "../../api/goal/Goals";
+import { IndependentOfEnvironment, ProductionEnvironment, StagingEnvironment } from "../../api/goal/support/environment";
 import { ArtifactGoal, AutofixGoal, BuildGoal, ReviewGoal, StagingEndpointGoal } from "../../api/machine/wellKnownGoals";
 import { StagingDeploymentGoal } from "../../api/machine/wellKnownGoals";
 import { DockerBuildGoal, TagGoal, VersionGoal } from "./commonGoals";
