+++ src/index.ts
@@ -24,12 +24,13 @@ export * from "./api/goal/ExecuteGoalResult";
 export * from "./api/goal/ExecuteGoalWithLog";
 export * from "./api/goal/Goal";
 export * from "./api/goal/Goals";
-export * from "./api/goal/SdmGoalImplementationMapper";
 export * from "./api/goal/common/GenericGoal";
 export * from "./api/goal/common/MessageGoal";
-export * from "./api/goal/common/approvalGate";
-export * from "./api/goal/common/environment";
-export * from "./api/goal/common/executeSendMessageToSlack";
+export * from "./api/goal/support/IsolatedGoalLauncher";
+export * from "./api/goal/support/SdmGoalImplementationMapper";
+export * from "./api/goal/support/approvalGate";
+export * from "./api/goal/support/environment";
+export * from "./api/goal/support/executeSendMessageToSlack";
 export * from "./api/listener/ArtifactListener";
 export * from "./api/listener/BuildListener";
 export * from "./api/listener/ChannelLinkListenerInvocation";
