+++ src/pack/freeze/deploymentFreeze.ts
@@ -2,8 +2,8 @@ import { logger } from "@atomist/automation-client";
 import { commandHandlerFrom } from "@atomist/automation-client/onCommand";
 import { EmptyParameters } from "../../api/command/EmptyParameters";
 import { allOf } from "../../api/dsl/allOf";
-import { executeSendMessageToSlack } from "../../api/goal/common/executeSendMessageToSlack";
 import { MessageGoal } from "../../api/goal/common/MessageGoal";
+import { executeSendMessageToSlack } from "../../api/goal/support/executeSendMessageToSlack";
 import { PushListenerInvocation } from "../../api/listener/PushListener";
 import { ExtensionPack } from "../../api/machine/ExtensionPack";
 import { PushTest } from "../../api/mapping/PushTest";
